Refactor EoMT segmentation transforms into a shared package
Move the instance/panoptic/semantic segmentation transform modules into a
new `_transforms/eomt_transforms/` package, prefix their classes with `EoMT`,
and factor the duplicated `resolve_incompatible` and transform-list-building
logic into a shared `utils.py`. Behavior is unchanged: the shared helpers
reproduce the previous per-module logic exactly (semantic still passes
`ignore_index` as `RandomCrop`'s `fill_mask`; instance/panoptic still don't).

Consumers in `_data/` and the task-model `transforms.py` subclasses are updated
to the new names; the concrete `DINOv2/v3EoMT*`, `Linear*`, and `Multihead*`
config classes keep their names, so existing configs/checkpoints are unaffected.
