+++ src/lightly_train/_data/mask_semantic_segmentation_dataset.py
@@ -23,10 +23,10 @@ from lightly_train._data.file_helpers import ImageMode
 from lightly_train._data.task_data_args import TaskDataArgs
 from lightly_train._data.task_dataset import TaskDataset, TaskDatasetArgs
 from lightly_train._env import Env
-from lightly_train._transforms.semantic_segmentation_transform import (
-    SemanticSegmentationCollateFunction,
-    SemanticSegmentationTransform,
-    SemanticSegmentationTransformArgs,
+from lightly_train._transforms.eomt_transforms import (
+    EoMTSemanticSegmentationCollateFunction,
+    EoMTSemanticSegmentationTransform,
+    EoMTSemanticSegmentationTransformArgs,
 )
 from lightly_train._transforms.task_transform import TaskCollateFunction, TaskTransform
 from lightly_train.types import (
@@ -63,14 +63,14 @@ class MaskSemanticSegmentationDataset(TaskDataset):
     dataset_args: MaskSemanticSegmentationDatasetArgs
 
     batch_collate_fn_cls: ClassVar[type[TaskCollateFunction]] = (
-        SemanticSegmentationCollateFunction
+        EoMTSemanticSegmentationCollateFunction
     )
 
     def __init__(
         self,
         dataset_args: MaskSemanticSegmentationDatasetArgs,
         image_info: Sequence[dict[str, str]],
-        transform: SemanticSegmentationTransform | None = None,
+        transform: EoMTSemanticSegmentationTransform | None = None,
     ):
         super().__init__(
             transform=transform, dataset_args=dataset_args, image_info=image_info
@@ -93,12 +93,12 @@ class MaskSemanticSegmentationDataset(TaskDataset):
 
     def set_transform(self, transform: TaskTransform) -> None:
         super().set_transform(transform)
-        assert isinstance(transform, SemanticSegmentationTransform)
+        assert isinstance(transform, EoMTSemanticSegmentationTransform)
         self._init_image_mode(transform)
 
-    def _init_image_mode(self, transform: SemanticSegmentationTransform) -> None:
+    def _init_image_mode(self, transform: EoMTSemanticSegmentationTransform) -> None:
         transform_args = transform.transform_args
-        assert isinstance(transform_args, SemanticSegmentationTransformArgs)
+        assert isinstance(transform_args, EoMTSemanticSegmentationTransformArgs)
 
         image_mode = (
             None
